requirements.txt
streamlit==1.32.0
pillow==10.2.0
opencv-python-headless==4.8.1.78
networkx==3.2.1
matplotlib==3.8.2
plotly==5.18.0
pandas==2.1.4
openai==1.12.0
python-dotenv==1.0.0
numpy==1.26.3
python-doctr==0.7.0
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
